Added Dummy data Added table

+++ app/src/main/java/com/example/sortay/data/CryptocurrencyData.kt
@@ -2,6 +2,6 @@ package com.example.sortay.data
 
 data class CryptocurrencyData(
      val name: String,
-     val price: Int,
-     val percentage: Int
+     val price: Double,
+     val percentage: Double
 )
